backend/requirements.txt
motor>=3.1.1
pymongo>=4.3.3
pydantic>=2.0.0
pydantic-settings
dnspython
numpy
scikit-learn
pandas
python-multipart
google-generativeai
python-dotenv
